+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/Transforms.java
@@ -62,6 +62,8 @@ final class Transforms {
     private static final Pattern NON_DIGIT_PATTERN = Pattern.compile("[^0-9]+");
     private static final float DEFAULT_CONFIDENCE_VALUE = 1.0f;
     private static final int DEFAULT_TABLE_SPAN = 1;
+    public static final String NORMALIZATION_ERROR_MESSAGE = "Please use the text property "
+        + "in case it was a normalization error";
 
     private Transforms() {
     }
@@ -350,25 +352,30 @@ final class Transforms {
                     FieldValueType.STRING);
                 break;
             case TIME:
-                LocalTime fieldTime = fieldValue.getValueTime() == null ? null : LocalTime
-                    .parse(fieldValue.getValueTime(), DateTimeFormatter.ofPattern("HH:mm:ss"));
+                if (fieldValue.getValueTime() != null) {
+                LocalTime fieldTime = LocalTime.parse(fieldValue.getValueTime(),
+                    DateTimeFormatter.ofPattern("HH:mm:ss"));
                 value = new com.azure.ai.formrecognizer.models.FieldValue(fieldTime, FieldValueType.TIME);
+                } else {
+                    throw LOGGER.logExceptionAsError(new RuntimeException(NORMALIZATION_ERROR_MESSAGE));
+                }
                 break;
             case DATE:
+                if (fieldValue.getValueDate() != null) {
                 value = new com.azure.ai.formrecognizer.models.FieldValue(fieldValue.getValueDate(),
                     FieldValueType.DATE);
+                } else {
+                    throw LOGGER.logExceptionAsError(new RuntimeException(NORMALIZATION_ERROR_MESSAGE));
+                }
                 break;
             case INTEGER:
-                com.azure.ai.formrecognizer.models.FieldValue longFieldValue;
-                if (fieldValue.getValueInteger() == null) {
-                    longFieldValue =
-                        new com.azure.ai.formrecognizer.models.FieldValue(null, FieldValueType.LONG);
-                } else {
-                    longFieldValue =
+                if (fieldValue.getValueInteger() != null) {
+                    value =
                         new com.azure.ai.formrecognizer.models.FieldValue(fieldValue.getValueInteger().longValue(),
                             FieldValueType.LONG);
+                } else {
+                    value = new com.azure.ai.formrecognizer.models.FieldValue(null, FieldValueType.LONG);
                 }
-                value = longFieldValue;
                 break;
             case NUMBER:
                 value = new com.azure.ai.formrecognizer.models.FieldValue(fieldValue.getValueNumber(),
@@ -380,21 +387,25 @@ final class Transforms {
                 break;
             case OBJECT:
                 value = new com.azure.ai.formrecognizer.models.FieldValue(
-                    toFieldValueObject(fieldValue.getValueObject(), readResults), FieldValueType.MAP);
+                        toFieldValueObject(fieldValue.getValueObject(), readResults), FieldValueType.MAP);
                 break;
             case SELECTION_MARK:
-                com.azure.ai.formrecognizer.models.SelectionMarkState selectionMarkState;
-                final FieldValueSelectionMark fieldValueSelectionMarkState = fieldValue.getValueSelectionMark();
-                if (FieldValueSelectionMark.SELECTED.equals(fieldValueSelectionMarkState)) {
-                    selectionMarkState = com.azure.ai.formrecognizer.models.SelectionMarkState.SELECTED;
-                } else if (FieldValueSelectionMark.UNSELECTED.equals(fieldValueSelectionMarkState)) {
-                    selectionMarkState = com.azure.ai.formrecognizer.models.SelectionMarkState.UNSELECTED;
+                if (fieldValue.getValueSelectionMark() != null) {
+                    com.azure.ai.formrecognizer.models.SelectionMarkState selectionMarkState;
+                    final FieldValueSelectionMark fieldValueSelectionMarkState = fieldValue.getValueSelectionMark();
+                    if (FieldValueSelectionMark.SELECTED.equals(fieldValueSelectionMarkState)) {
+                        selectionMarkState = com.azure.ai.formrecognizer.models.SelectionMarkState.SELECTED;
+                    } else if (FieldValueSelectionMark.UNSELECTED.equals(fieldValueSelectionMarkState)) {
+                        selectionMarkState = com.azure.ai.formrecognizer.models.SelectionMarkState.UNSELECTED;
+                    } else {
+                        selectionMarkState = com.azure.ai.formrecognizer.models.SelectionMarkState.fromString(
+                            fieldValue.getValueSelectionMark().toString());
+                    }
+                    value = new com.azure.ai.formrecognizer.models.FieldValue(selectionMarkState,
+                        FieldValueType.SELECTION_MARK_STATE);
                 } else {
-                    selectionMarkState = com.azure.ai.formrecognizer.models.SelectionMarkState.fromString(
-                        fieldValue.getValueSelectionMark().toString());
+                    throw LOGGER.logExceptionAsError(new RuntimeException(NORMALIZATION_ERROR_MESSAGE));
                 }
-                value = new com.azure.ai.formrecognizer.models.FieldValue(selectionMarkState,
-                    FieldValueType.SELECTION_MARK_STATE);
                 break;
             case COUNTRY_REGION:
                 value = new com.azure.ai.formrecognizer.models.FieldValue(fieldValue.getValueCountryRegion(),
@@ -430,6 +441,9 @@ final class Transforms {
      */
     private static Map<String, FormField> toFieldValueObject(Map<String, FieldValue> valueObject,
                                                              List<ReadResult> readResults) {
+        if (valueObject == null) {
+            return null;
+        }
         Map<String, FormField> fieldValueObjectMap = new TreeMap<>();
         valueObject.forEach((key, fieldValue) -> {
 
@@ -456,6 +470,9 @@ final class Transforms {
      * @return The List of {@link FormField}.
      */
     private static List<FormField> toFieldValueArray(List<FieldValue> valueArray, List<ReadResult> readResults) {
+        if (valueArray == null) {
+            return null;
+        }
         return valueArray.stream()
             .map(fieldValue -> {
                 FieldData valueData = null;