+++ sdk/formrecognizer/azure-ai-formrecognizer/src/test/java/com/azure/ai/formrecognizer/FormRecognizerClientTest.java
@@ -5,6 +5,7 @@ package com.azure.ai.formrecognizer;
 
 import com.azure.ai.formrecognizer.models.CreateComposedModelOptions;
 import com.azure.ai.formrecognizer.models.FormContentType;
+import com.azure.ai.formrecognizer.models.FormField;
 import com.azure.ai.formrecognizer.models.FormPage;
 import com.azure.ai.formrecognizer.models.FormRecognizerErrorInformation;
 import com.azure.ai.formrecognizer.models.FormRecognizerException;
@@ -35,6 +36,7 @@ import java.io.InputStream;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 import static com.azure.ai.formrecognizer.FormRecognizerClientTestBase.PrebuiltType.BUSINESS_CARD;
 import static com.azure.ai.formrecognizer.FormRecognizerClientTestBase.PrebuiltType.IDENTITY;
@@ -2048,6 +2050,32 @@ public class FormRecognizerClientTest extends FormRecognizerClientTestBase {
                 Context.NONE).setPollInterval(durationTestMode);
             validateNetworkCallRecord("locale", "en-US");
         }, INVOICE_PDF);
+    }
+
+    /**
+     * Verify SDK returns empty object and array for null sub line items field.
+     */
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.formrecognizer.TestUtils#getTestParameters")
+    public void invoiceSubLineItemsNull(HttpClient httpClient, FormRecognizerServiceVersion serviceVersion) {
+        client = getFormRecognizerClient(httpClient, serviceVersion);
+        localFilePathRunner((filePath, dataLength) -> {
+            List<RecognizedForm> recognizedForms = client.beginRecognizeInvoices(
+                getContentDetectionFileData(filePath),
+                dataLength,
+                new RecognizeInvoicesOptions().setLocale(FormRecognizerLocale.EN_US),
+                Context.NONE)
+                .setPollInterval(durationTestMode)
+                .getFinalResult();
+
+            RecognizedForm recognizedForm = recognizedForms.get(0);
+            FormField itemFieldList = recognizedForm.getFields().get("Items").getValue().asList().get(0);
+            Map<String, FormField> formFieldMap = itemFieldList.getValue().asMap();
+
+            assertEquals(null, formFieldMap);
+            assertEquals(String.valueOf(1), itemFieldList.getValueData().getText());
+
+        }, INVOICE_NO_SUB_LINE_PDF);
     }
 
     // Identity Document Recognition